Move `getArrivalDate` to its own file
This prevents dependency cylces. I’ve also renamed it to be clearer
about what exactly it is doing.
