+++ server/utils/applications/arrivalDateFromApplication.ts
@@ -1,0 +1,48 @@
+import type { ApprovedPremisesApplication as Application } from '@approved-premises/api'
+import { SessionDataError } from '../errors'
+
+export const arrivalDateFromApplication = (application: Application, raiseOnMissing = true): string | null => {
+  const throwOrReturnNull = (message: string): null => {
+    if (raiseOnMissing) {
+      throw new SessionDataError(message)
+    }
+
+    return null
+  }
+
+  const basicInformation = application.data?.['basic-information']
+
+  if (!basicInformation) return throwOrReturnNull('No basic information')
+
+  const {
+    knowReleaseDate = '',
+    startDateSameAsReleaseDate = '',
+    releaseDate = '',
+    startDate = '',
+  } = {
+    ...basicInformation['release-date'],
+    ...basicInformation['placement-date'],
+  }
+
+  if (!knowReleaseDate || knowReleaseDate === 'no') {
+    return throwOrReturnNull('No known release date')
+  }
+
+  if (knowReleaseDate === 'yes' && startDateSameAsReleaseDate === 'yes') {
+    if (!releaseDate) {
+      return throwOrReturnNull('No release date')
+    }
+
+    return releaseDate
+  }
+
+  if (startDateSameAsReleaseDate === 'no') {
+    if (!startDate) {
+      return throwOrReturnNull('No start date')
+    }
+
+    return startDate
+  }
+
+  return null
+}
